[First Commit] Middleware + Demo
Introducing a middleware for Gin Gonic that forwards request logs to
a fluentd HTTP server (hooked to an ElasticSearch/Kibana stack for
instance).

It supports different options (logging bodies or not - or just on
errors, limiting the number of buffered request to be forwarded to
fluentd to mitigate DDoS attacks, limiting the size of logged
bodies...).