[v1.0.0] Now this is prod ready for real
After @jindav's review, a lot was changed. The program has been
completely refactored. We use io golang's standard library
abstractions to limit the memory overhead of the middleware as much
as possible.

The POST to fluentd is now retried. In case of failure to connect to
Fluentd, the memory consumption is limited by the size of the channel.

Over all, the code is much more idiomatic.